+++ scripts/timer-planning.js
@@ -5,7 +5,7 @@
 // -- Import MODULE variables --
 import { MODULE, BLACKSMITH } from './const.js';
 import { COFFEEPUB, postConsoleAndNotification, playSound, trimString } from './api-common.js';
-import { CombatStats } from './combat-stats.js';
+import { CombatStats } from './stats-combat.js';
 import { SocketManager } from './manager-sockets.js';
 
 export class PlanningTimer {
@@ -735,7 +735,7 @@ export class PlanningTimer {
             ...data
         };
 
-        const messageHtml = await renderTemplate('modules/coffee-pub-blacksmith/templates/chat-cards.hbs', messageData);
+        const messageHtml = await renderTemplate('modules/coffee-pub-blacksmith/templates/cards-common.hbs', messageData);
 
         await ChatMessage.create({
             content: messageHtml,