+++ scripts/token-movement.js
@@ -320,7 +320,7 @@ export class MovementConfig extends Application {
                 movementDescription: movementType.description
             };
 
-            const basicContent = await renderTemplate('modules/coffee-pub-blacksmith/templates/chat-cards.hbs', basicTemplateData);
+            const basicContent = await renderTemplate('modules/coffee-pub-blacksmith/templates/cards-common.hbs', basicTemplateData);
 
             // Send chat message
             ChatMessage.create({
@@ -1067,7 +1067,7 @@ async function postMarchingOrder() {
     const spacingText = spacing > 0 ? `${spacing} grid space${spacing > 1 ? 's' : ''}` : '';
 
     // Create chat message
-    const content = await renderTemplate('modules/coffee-pub-blacksmith/templates/chat-cards.hbs', {
+    const content = await renderTemplate('modules/coffee-pub-blacksmith/templates/cards-common.hbs', {
         isPublic: true,
         isMovementChange: true,
         movementIcon: movementType.icon,
@@ -1153,7 +1153,7 @@ Hooks.on('createCombat', async (combat) => {
                 movementDescription: `When combat ends <strong>${prevModeType.name} Mode</strong> will be restored.<br><br>${combatModeType.description}`
             };
 
-            const combatContent = await renderTemplate('modules/coffee-pub-blacksmith/templates/chat-cards.hbs', combatTemplateData);
+            const combatContent = await renderTemplate('modules/coffee-pub-blacksmith/templates/cards-common.hbs', combatTemplateData);
 
             ChatMessage.create({
                 content: combatContent,
@@ -1204,7 +1204,7 @@ Hooks.on('deleteCombat', async (combat) => {
             movementDescription: `${movementType.name} Mode restored.<br><br>${movementType.description}`
         };
 
-        const endCombatContent = await renderTemplate('modules/coffee-pub-blacksmith/templates/chat-cards.hbs', endCombatTemplateData);
+        const endCombatContent = await renderTemplate('modules/coffee-pub-blacksmith/templates/cards-common.hbs', endCombatTemplateData);
 
         ChatMessage.create({
             content: endCombatContent,